+++ testament/tester.nim
@@ -12,7 +12,9 @@
 import
   parseutils, strutils, pegs, os, osproc, streams, parsecfg, json,
   marshal, backend, parseopt, specs, htmlgen, browsers, terminal,
-  algorithm, compiler/nodejs, times, sets, md5, sequtils
+  algorithm, times, sets, md5, sequtils
+
+include compiler/nodejs
 
 var useColors = true
 var backendLogging = true